# jaxrs-spec-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API spec for DCFB API.",
    "version" : "0.0.1",
    "title" : "DCFB API"
  },
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json;charset=utf-8" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "security" : [ {
    "bearer" : [ ]
  } ],
  "paths" : {
...
    "/purchaseRequests" : {
      "get" : {
        "tags" : [ "PurchaseRequests" ],
        "summary" : "List purchase requests",
        "description" : "List purchase requests",
        "operationId" : "list purchase requests",
        "parameters" : [ {
          "name" : "categoryIds",
          "in" : "query",
          "description" : "Filter by category ids (comma delimitered)",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "search",
          "in" : "query",
          "description" : "Search  by free-text query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "An array of purchase requests",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PurchaseRequest"
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "PurchaseRequests" ],
        "summary" : "Create purchase request",
        "description" : "Create purchase request",
        "operationId" : "createPurchaseRequest",
        "parameters" : [ {
          "in" : "body",
          "name" : "payload",
          "description" : "Payload",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/PurchaseRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Created purchaseRequest",
            "schema" : {
              "$ref" : "#/definitions/PurchaseRequest"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/purchaseRequests/{purchaseRequestId}" : {
      "get" : {
        "tags" : [ "PurchaseRequests" ],
        "summary" : "Finds an purchaseRequest",
        "description" : "Find purchaseRequest",
        "operationId" : "findPurchaseRequest",
        "parameters" : [ {
          "name" : "purchaseRequestId",
          "in" : "path",
          "description" : "PurchaseRequest id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "An array of purchase requests",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PurchaseRequest"
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "PurchaseRequests" ],
        "summary" : "updates an purchase request",
        "description" : "Update purchase request",
        "operationId" : "updatePurchaseRequest",
        "parameters" : [ {
          "name" : "purchaseRequestId",
          "in" : "path",
          "description" : "PurchaseRequest id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "payload",
          "description" : "Payload",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/PurchaseRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Updated purchaseRequest",
            "schema" : {
              "$ref" : "#/definitions/PurchaseRequest"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "PurchaseRequests" ],
        "summary" : "deletes an purchase request",
        "description" : "Delete purchase request",
        "operationId" : "deletePurchaseRequest",
        "parameters" : [ {
          "name" : "purchaseRequestId",
          "in" : "path",
          "description" : "PurchaseRequest id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "204" : {
            "description" : "Empty resonse indicating a succesfull removal"
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "bearer" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "LocalizedValue" : {
      "type" : "object",
      "required" : [ "language", "type", "value" ],
      "properties" : {
        "language" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
...
    "PurchaseRequest" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Purchase request id"
        },
        "title" : {
          "type" : "array",
          "description" : "Title of the item, multilingual",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "description" : {
          "type" : "array",
          "description" : "Description of the item, multilingual",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "categoryId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Category id."
        },
        "slug" : {
          "type" : "string",
          "description" : "Item slug"
        },
        "createdAt" : {
          "type" : "string",
          "format" : "date-time"
        },
        "modifiedAt" : {
          "type" : "string",
          "format" : "date-time"
        },
        "expiresAt" : {
          "type" : "string",
          "format" : "date-time"
        },
        "unit" : {
          "type" : "string"
        },
        "amount" : {
          "type" : "number",
          "format" : "int64"
        }
      }
    },
    "InternalServerError" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "Forbidden" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "BadRequest" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
...
  }
}